big update

+++ app/controllers/PostsController.php
@@ -10,7 +10,11 @@ class PostsController extends \BaseController {
 	public function index()
 	{
 		$posts = Post::paginate(5);
-		return View::make('index')->with('posts', $posts)->with('numPosts',count($posts));
+		$tags = array();
+		foreach ($posts as $post) {
+			$tags[] = $post->tags()->get()->lists('tag_text');
+		}
+		return View::make('index')->withPosts($posts)->with('numPosts',count($posts))->withTags($tags);
 	}
 
 
@@ -75,12 +79,13 @@ class PostsController extends \BaseController {
 	public function show($id)
 	{
 		$post = Post::find($id);
+		$tags = $post->tags()->get()->lists('tag_text');
 		if(isset($post)) 
 		{
 			$post->visits_counter++;
 			$post->save();
 		}
-		return View::make('posts.show')->withPost($post);
+		return View::make('posts.show')->withPost($post)->withTags($tags);
 	}
 
 
@@ -92,39 +97,44 @@ class PostsController extends \BaseController {
 	 */
 	public function edit($id)
 	{
-		$rules = array
-		(
-			'title' => 'required|min:3',
-			'text' => 'required|min:10',
-			'tags' => 'required'
-		);
-
-		$validator = Validator::make(Input::all(), $rules);
-
-		if($validator->fails())
+		if(Auth::check() && Auth::user()->user_level == '1') 
 		{
-			return Redirect::to('/post/edit/' . $id);
+			$rules = array
+			(
+				'title' => 'required|min:3',
+				'text' => 'required|min:10',
+				'tags' => 'required'
+			);
+
+			$validator = Validator::make(Input::all(), $rules);
+
+			if($validator->fails())
+			{
+				return Redirect::to('/post/edit/' . $id);
+			}
+			else
+			{
+				$post = Post::find($id);
+				$post->post_title = Input::get('title');
+				$post->post_text = Input::get('text');
+				$post->save();
+				foreach ($post->tags()->get() as $key => $tag) {
+					$tag->delete();
+				}
+				$tags = explode(',', Input::get('tags'));
+				foreach($tags as $value)
+				{
+					$value = trim($value);
+					$tag = new Tag;
+					$tag->post_id = $post->id;
+					$tag->tag_text = $value;
+					$tag->save();
+				}
+				return Redirect::to('/');
+			}
 		}
 		else
-		{
-			$post = Post::find($id);
-			$post->post_title = Input::get('title');
-			$post->post_text = Input::get('text');
-			$post->save();
-			foreach ($post->tags()->get() as $key => $tag) {
-				$tag->delete();
-			}
-			$tags = explode(',', Input::get('tags'));
-			foreach($tags as $value)
-			{
-				$value = trim($value);
-				$tag = new Tag;
-				$tag->post_id = $post->id;
-				$tag->tag_text = $value;
-				$tag->save();
-			}
-			return Redirect::to('/');
-		}
+			return View::make('pages.pageNotFound');
 	}
 
 
@@ -148,19 +158,22 @@ class PostsController extends \BaseController {
 	 */
 	public function destroy($id)
 	{
-		$post = Post::find($id);
-		$tags = $post->tags()->get();
-		$comments = $post->comments()->get();
+		if(Auth::check() && Auth::user()->user_level == '1') 
+		{
+			$post = Post::find($id);
+			$tags = $post->tags()->get();
+			$comments = $post->comments()->get();
 
-		foreach ($comments as $key => $comment) {
-			$comment->delete();
+			foreach ($comments as $key => $comment) {
+				$comment->delete();
+			}
+			foreach ($tags as $key => $tag) {
+				$tag->delete();
+			}
+			$post->delete();
+			return Redirect::to('/');
 		}
-		foreach ($tags as $key => $tag) {
-			$tag->delete();
-		}
-		$post->delete();
-		return Redirect::to('/');
+		else
+			return View::make('pages.pageNotFound');
 	}
-
-
 }
